mnist-digit-recognizer/requirements.txt
scikit-learn
numpy
matplotlib
gradio
pillow
joblib
